Added feature: snap points
Snap points are trigger volumes which reposition interactables to target positions, when interaction on the interactable has stopped while clipping with the trigger volume.

Snap points take into account whether it is dependent on other Snap points to already be hold objects.
Snap points take into account if an object has to right interatable tag before performing a snap.
In the event the interactable clips through more than 1 valid Snap point, the closest snap point is chosen to perform the snap.
Once an interactable is snapped, it's rigidbody will be set to Kinematic, where it cannot be moved until interacted with again.