Added feature: focus point
A Focus point toggles between moving the player camera to a fixed position, and reverting back to normal.

In the scene, the desks are Focus points. A new interactable, Button which performs its action on click, activates the toggling of the Focus point.

When the desk is first clicked, the Focus point detaches the player from the moving collider, while also blocking movement controls, and attaches the player to the fixed immovable point.

When the desk is clicked again, the Focus point detaches the player from immovable point, and returns the player to the moving collider.