Added feature: surface normal indicator, save player settings
While an object has been grabbed, the surface normal indicator will appear.
The surface normal indicator will indicate the normal of the surface the object will be placed on when let go.

Settings menu now saves the player's mouse sensitivity when set.
The change will be reflected when the game is opened again.

Did some housekeeping on the code.