Added feature: free-hold mode
Added the ability to switch between the current object holding state referred to as "placement mode", and the newly added "free-hold mode". Switching the holding state is toggled by clicking the F key.

While in free-hold mode, the object being held can be freely brought closer or further with the scroll wheel.